Revert "Merge branch 'dev' into update-personal-info"
This reverts commit d775a268d98a71c8508fa9b94a7742aea1ab551b, reversing
changes made to e3f87319d73985247b4d111f6871909ff848b313.
