+++ .gitignore
@@ -22,5 +22,3 @@
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
 replay_pid*
-src/.DS_Store
-src/.DS_Store
