+++ README.md
@@ -17,45 +17,38 @@
 6. A bank customer should be able to close an existing account.
 7. A bank customer should be able to transfer money from one account to another. 
 8. A bank adminstrator should be able to collect fees from existing accounts when necessary.
-9. A bank adminstrator should be able to add an interest payment to an existing account when necessary.、
-10. A bank administrator should be able to login as a administrator to perform actions that only a bank admin can perform.
-11. A bank customer should be able to receive notifications when they have fee before due.
-12. A bank administrator should be able to see all transactions for a particular account.
-13. A bank customer should be able to set a auto recurring payment to an account.
-14. A bank administrator should be able to freeze or unfreeze a customer account.
-15. A bank costumer should be able to set a maximum withdraw amount for their account
-16. A bank customer should be able to set a password or PIN for account security.
-17. A bank customer should be able to update their personal information.
+9. A bank adminstrator should be able to add an interest payment to an existing account when necessary.
 
 ## What user stories were completed in this iteration?
-* In this iteration, we completed several user stories related to both customer and administrator functions. A bank administrator can now log in as an administrator, view all transactions for a particular account, and freeze or unfreeze a customer account. On the customer side, users can now receive notifications before a fee is due, set up an auto recurring payment to an account, set a maximum withdrawal amount, create a password or PIN for account security, and update their personal information.
+* In this iteration, our team completed the user stories that allow a bank customer to withdraw from an account, check their account balance, view transaction history for an account, create an additional account with the bank, close an existing account, and transfer money from one account to another, as well as the user story that allows a bank administrator to add an interest payment to an existing account when necessary.
 
 ## What user stories do you intend to complete next iteration?
-In the next iteration, we plan to improve and expand the current banking features by adding more account management options, improving transaction handling, and strengthening security-related functions. We also want to continue refining the menu structure and overall usability of the program.
+1. A bank customer should be able to update their personal information 
+2. A bank customer should be able to receive notifications for important account activities.
+3. A bank customer should be able to set a password or PIN for account security.
+4. A bank administrator should be able to freeze or unfreeze a customer account.
+5. The system should log all transactions for auditing purposes.
+6. The bank should be able have a list of account associated with a customer. 
 
 ## What user stories were completed in this iteration?
-1. A bank administrator should be able to login as a administrator to perform actions that only a bank admin can perform.
-2. A bank customer should be able to receive notifications when they have fee before due.
-3. A bank administrator should be able to see all transactions for a particular account.
-4. A bank customer should be able to set a auto recurring payment to an account.
-5. A bank administrator should be able to freeze or unfreeze a customer account.
-6. A bank costumer should be able to set a maximum withdraw amount for their account
-7. A bank customer should be able to set a password or PIN for account security.
-8. A bank customer should be able to update their personal information.
+1. A bank customer should be able to withdraw from an account. 
+2. A bank customer should be able to check their account balance. 
+3. A bank customer should be able to view their transaction history for an account. 
+4. A bank customer should be able to create an additional account with the bank. 
+5. A bank customer should be able to close an existing account.
+6. A bank customer should be able to transfer money from one account to another. 
+7. A bank adminstrator should be able to collect fees from existing accounts when necessary.
+8. A bank adminstrator should be able to add an interest payment to an existing account when necessary.
+
 
 
 ## Is there anything that you implemented but doesn't currently work?
-* At this point, the main features listed above have been implemented. Some parts may still need further testing, integration, or refinement to make sure they work smoothly in all cases, but no major intended feature is completely unfinished.
+* All functionalities work properly, however, we may need to create a class called bank to host open and close account logic. 
 
 
 ## What commands are needed to compile and run your code from the command line?
 ```bash
 chmod +x runApp.sh
 ./runApp.sh
-```
-
-## Admin Password
-```text
-admin123
 ```
