+++ src/main/BankAccount.java
@@ -1,8 +1,6 @@
 package main;
 
 import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.List;
 
 public class BankAccount implements Serializable {
 
@@ -10,81 +8,44 @@ public class BankAccount implements Serializable {
 
     private double balance;
     private String transactionHistory;
-    private ArrayList<Fee> fees;
-    private boolean frozen;
-    private double maxWithdrawAmount;
 
     public BankAccount() {
         this.balance = 0;
         this.transactionHistory = "";
-        this.fees = new ArrayList<>();
-        this.frozen = false;
-        this.maxWithdrawAmount = Double.MAX_VALUE;
     }
 
     public BankAccount(double balance, String transactionHistory) {
         this.balance = balance;
         this.transactionHistory = transactionHistory == null ? "" : transactionHistory;
-        this.fees = new ArrayList<>();
-        this.frozen = false;
-        this.maxWithdrawAmount = Double.MAX_VALUE;
-    }
-
-    public BankAccount(double balance, String transactionHistory, boolean frozen) {
-        this.balance = balance;
-        this.transactionHistory = transactionHistory == null ? "" : transactionHistory;
-        this.fees = new ArrayList<>();
-        this.frozen = frozen;
-        this.maxWithdrawAmount = Double.MAX_VALUE;
-    }
-
-    public BankAccount(double balance, String transactionHistory, boolean frozen, double maxWithdrawAmount) {
-        this.balance = balance;
-        this.transactionHistory = transactionHistory == null ? "" : transactionHistory;
-        this.fees = new ArrayList<>();
-        this.frozen = frozen;
-        this.maxWithdrawAmount = maxWithdrawAmount <= 0 ? Double.MAX_VALUE : maxWithdrawAmount;
-    }
-
-    public BankAccount(double balance, String transactionHistory, double maxWithdrawAmount) {
-        this.balance = balance;
-        this.transactionHistory = transactionHistory == null ? "" : transactionHistory;
-        this.fees = new ArrayList<>();
-        this.frozen = false;
-        if (maxWithdrawAmount <= 0) {
-            this.maxWithdrawAmount = Double.MAX_VALUE;
-        } else {
-            this.maxWithdrawAmount = maxWithdrawAmount;
-        }
     }
 
     public void deposit(double amount) {
-        ensureAccountIsActive();
-        if (amount > 0) {
+        if(amount > 0) {
             this.balance += amount;
             this.transactionHistory += "Deposited: " + amount + "\n";
         } else {
-            throw new IllegalArgumentException("Deposit amount must be greater than 0.");
+            throw new IllegalArgumentException();
         }
     }
-
+    
     public void addInterest(double interestAmount) {
         if (interestAmount > 0) {
             this.balance += interestAmount;
             this.transactionHistory += "Interest added: " + interestAmount + "\n";
         } else {
-            throw new IllegalArgumentException("Interest amount must be greater than 0.");
+            throw new IllegalArgumentException();
         }
     }
 
+    //collect fee from the account, this will be used by the bank to collect fees for transactions
     public void collectFee(double feeAmount) {
-        if (feeAmount > 0) {
-            this.balance -= feeAmount;
-            this.transactionHistory += "Fee collected: " + feeAmount + "\n";
-        } else {
-            throw new IllegalArgumentException("Fee amount must be greater than 0.");
-        }
+    if (feeAmount > 0) {
+        this.balance -= feeAmount;
+        this.transactionHistory += "Fee collected: " + feeAmount + "\n";
+    } else {
+        throw new IllegalArgumentException();
     }
+}
 
     public double getBalance() {
         return this.balance;
@@ -93,64 +54,20 @@ public class BankAccount implements Serializable {
     public String getTransactionHistory() {
         return this.transactionHistory;
     }
-
-    public boolean isFrozen() {
-        return this.frozen;
-    }
-
-    public void freezeAccount() {
-        if (this.frozen) {
-            throw new IllegalStateException("This account is already frozen.");
-        }
-        this.frozen = true;
-        this.transactionHistory += "Account frozen.\n";
-    }
-
-    public void unfreezeAccount() {
-        if (!this.frozen) {
-            throw new IllegalStateException("This account is not frozen.");
-        }
-        this.frozen = false;
-        this.transactionHistory += "Account unfrozen.\n";
-    }
-
-    public double getMaxWithdrawAmount() {
-        return this.maxWithdrawAmount;
-    }
-
-    public void setMaxWithdrawAmount(double maxWithdrawAmount) {
-        if (maxWithdrawAmount <= 0) {
-            throw new IllegalArgumentException("Maximum withdrawal amount must be greater than 0.");
-        }
-
-        this.maxWithdrawAmount = maxWithdrawAmount;
-        this.transactionHistory += "Maximum withdrawal amount set to: " + maxWithdrawAmount + "\n";
-    }
+    //withdraw
     public void withdraw(double amount) {
-        ensureAccountIsActive();
-
         if (amount <= 0) {
             throw new IllegalArgumentException("Withdrawal amount must be greater than 0.");
-        }
-
-        if (amount > maxWithdrawAmount) {
-            throw new IllegalArgumentException("Withdrawal amount exceeds the maximum amount");
-        }
-        if (amount > balance) {
-            throw new IllegalArgumentException("Insufficient funds.");
+        }if (amount > balance) {
+            throw new IllegalArgumentException("You are overdrafting your account.");
         }
         balance -= amount;
         this.transactionHistory += "Withdrew: " + amount + "\n";
     }
 
     public void transferMoney(BankAccount targetAccount, double amount) {
-        ensureAccountIsActive();
-
         if (targetAccount == null) {
             throw new IllegalArgumentException("Source and target accounts cannot be null.");
-        }
-        if (targetAccount.isFrozen()) {
-            throw new IllegalStateException("You cannot transfer money into a frozen account.");
         }
         if (amount <= 0) {
             throw new IllegalArgumentException("Transfer amount must be positive.");
@@ -158,24 +75,7 @@ public class BankAccount implements Serializable {
         if (this.balance < amount) {
             throw new IllegalArgumentException("Insufficient funds in source account.");
         }
-
         this.withdraw(amount);
         targetAccount.deposit(amount);
-    }
-
-    public void createFee(Fee fee) {
-        if (fee == null) {
-            throw new IllegalArgumentException("Fee cannot be null.");
-        }
-        this.fees.add(fee);
-    }
-
-    public List<Fee> getRemainingFees() {
-        return new ArrayList<>(this.fees);
-    }
-    private void ensureAccountIsActive() {
-        if (this.frozen) {
-            throw new IllegalStateException("This account is frozen. Transactions are not allowed.");
-        }
     }
 }
