+++ src/main/MainMenu.java
@@ -5,17 +5,21 @@ import java.util.Scanner;
 
 public class MainMenu {
 
-    private static final int EXIT_SELECTION = 12;
-    private static final int MAX_SELECTION = 12;
+    private static final int EXIT_SELECTION = 11;
+    private static final int MAX_SELECTION = 11;
 
     private final Scanner keyboardInput;
+    private final Customer customer;
     private final Bank bank;
 
     public MainMenu() {
         this.keyboardInput = new Scanner(System.in);
         this.bank = BankDataStore.loadBank();
         if (this.bank.getCustomers().isEmpty()) {
-            this.bank.addCustomer(new Customer("Default User"));
+            this.customer = new Customer("Default User");
+            this.bank.addCustomer(this.customer);
+        } else {
+            this.customer = this.bank.getCustomers().get(0);
         }
     }
 
@@ -31,8 +35,7 @@ public class MainMenu {
         System.out.println("8. Admin: Collect fee");
         System.out.println("9. Admin: Add interest payment");
         System.out.println("10. Update personal information");
-        System.out.println("11. Set password or PIN");
-        System.out.println("12. Exit the app");
+        System.out.println("11. Exit the app");
     }
 
     public int getUserSelection(int max) {
@@ -52,10 +55,10 @@ public class MainMenu {
         try {
             switch (selection) {
                 case 1:
-                    runCustomerMenu();
+                    performDeposit();
                     break;
                 case 2:
-                    runAdminMenu();
+                    performWithdrawal();
                     break;
                 case 3:
                     displayBalance();
@@ -80,17 +83,39 @@ public class MainMenu {
                     break;
                 case 10:
                     updatePersonalInformation();
-                    setPasswordOrPin();
                     break;
                 case 11:
                     System.out.println("Thank you for using the 237 Bank App!");
                     break;
-                default:
-                    System.out.println("Invalid selection.");
-                    break;
             }
         } catch (IllegalStateException | IllegalArgumentException e) {
             System.out.println(e.getMessage());
+        }
+    }
+
+    public void performDeposit() {
+        BankAccount account = selectAccount("deposit into");
+        double depositAmount = readPositiveAmount("How much would you like to deposit: ");
+        account.deposit(depositAmount);
+        System.out.println("Deposit successful.");
+    }
+
+    public void displayTransactionHistory() {
+        BankAccount account = selectAccount("view transaction history for");
+        System.out.println("Transaction History:");
+        String history = account.getTransactionHistory();
+        System.out.println(history.isEmpty() ? "No transactions yet." : history);
+    }
+    
+    public void performWithdrawal() {
+        BankAccount account = selectAccount("withdraw from");
+        double withdrawAmount = readPositiveAmount("How much would you like to withdraw: ");
+
+        try {
+            account.withdraw(withdrawAmount);
+            System.out.println("Withdrawal successful.");
+        } catch (IllegalArgumentException e) {
+            System.out.println("Invalid withdrawal. Please make sure the amount is greater than 0 and does not exceed your balance.");
         }
     }
 
@@ -104,60 +129,60 @@ public class MainMenu {
         saveData();
     }
 
-    private void runCustomerMenu() {
-        Customer selectedCustomer = selectCustomer();
-        new CustomerMenu(bank, keyboardInput, selectedCustomer).run();
+    public void displayBalance() {
+        BankAccount account = selectAccount("check balance for");
+        System.out.println("Your current balance is: " + account.getBalance());
     }
 
-    private void runAdminMenu() {
-        System.out.print("Enter admin password: ");
-        String password = keyboardInput.next();
-        if (!AdminAuth.isValidPassword(password)) {
-            System.out.println("Invalid admin password.");
-            return;
-        }
-        Customer defaultCustomer = bank.getCustomers().get(0);
-        new AdminMenu(bank, keyboardInput, defaultCustomer).run();
+    public void createAdditionalAccount() {
+        customer.openAccount();
+        System.out.println("Account created successfully.");
     }
 
-    private Customer selectCustomer() {
-        System.out.println("Select customer:");
-        for (int i = 0; i < bank.getCustomers().size(); i++) {
-            System.out.println((i + 1) + ". " + bank.getCustomers().get(i).getName());
+    public void closeExistingAccount() {
+        BankAccount account = selectAccount("close");
+        try {
+            customer.closeAccount(account);
+            System.out.println("Account closed successfully.");
+        } catch (IllegalStateException e) {
+            System.out.println(e.getMessage());
         }
-        int selectedCustomer = getUserSelection(bank.getCustomers().size());
-        return bank.getCustomers().get(selectedCustomer - 1);
     }
 
-    public void viewAdminTransactionHistory() {
-        List<Customer> customers = bank.getCustomers();
-        if (customers.isEmpty()) {
-            System.out.println("No customers found.");
-            return;
-        }
-        System.out.println("Select a customer:");
-        for (int i = 0; i < customers.size(); i++) {
-            System.out.println((i + 1) + ". " + customers.get(i).getName());
-        }
-        Customer selected = customers.get(getUserSelection(customers.size()) - 1);
+    public void transferBetweenAccounts() {
+        System.out.println("Select source account:");
+        BankAccount sourceAccount = selectAccount("transfer money from");
+        System.out.println("Select target account:");
+        BankAccount targetAccount = selectAccount("transfer money to");
+        double amount = readPositiveAmount("How much would you like to transfer: ");
 
-    public void setPasswordOrPin() {
-        keyboardInput.nextLine();
-        System.out.println("1. Set password");
-        System.out.println("2. Set PIN");
-        int choice = getUserSelection(2);
-        keyboardInput.nextLine();
+        try {
+            sourceAccount.transferMoney(targetAccount, amount);
+            System.out.println("Transfer successful.");
+        } catch (IllegalArgumentException e) {
+            System.out.println(e.getMessage());
+        }
+    }
 
-        if (choice == 1) {
-            System.out.print("Enter your new password: ");
-            String password = keyboardInput.nextLine();
-            customer.setPassword(password);
-            System.out.println("Password set successfully.");
-        } else {
-            System.out.print("Enter your 4-digit PIN: ");
-            String pin = keyboardInput.nextLine();
-            customer.setPin(pin);
-            System.out.println("PIN set successfully.");
+    public void collectFee() {
+        BankAccount account = selectAccount("collect fee from");
+        double feeAmount = readPositiveAmount("Enter fee amount: ");
+        try {
+            bank.collectFee(account, feeAmount);
+            System.out.println("Fee collected successfully.");
+        } catch (IllegalArgumentException e) {
+            System.out.println("Invalid fee amount.");
+        }
+    }
+
+    public void addInterest() {
+        BankAccount account = selectAccount("add interest to");
+        double interestAmount = readPositiveAmount("Enter interest amount: ");
+        try {
+            bank.addInterest(account, interestAmount);
+            System.out.println("Interest added successfully.");
+        } catch (IllegalArgumentException e) {
+            System.out.println("Invalid interest amount.");
         }
     }
 
@@ -180,11 +205,21 @@ public class MainMenu {
         for (int i = 0; i < accounts.size(); i++) {
             System.out.println((i + 1) + ". Account #" + (i + 1) + " (Balance: " + accounts.get(i).getBalance() + ")");
         }
-        BankAccount account = accounts.get(getUserSelection(accounts.size()) - 1);
+        int selectedAccount = getUserSelection(accounts.size());
+        return accounts.get(selectedAccount - 1);
+    }
 
-        String history = account.getTransactionHistory();
-        System.out.println("Transaction History:");
-        System.out.println(history.isEmpty() ? "No transactions yet." : history);
+    private double readPositiveAmount(String prompt) {
+        double amount = -1;
+        while (amount <= 0) {
+            System.out.print(prompt);
+            if (keyboardInput.hasNextDouble()) {
+                amount = keyboardInput.nextDouble();
+            } else {
+                keyboardInput.next();
+            }
+        }
+        return amount;
     }
 
     private String readRequiredText(String prompt) {
