+++ src/test/BankTest.java
@@ -6,8 +6,6 @@ import main.Customer;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.assertFalse;
 
 public class BankTest {
 
@@ -37,51 +35,5 @@ public class BankTest {
         account.deposit(100.0);
         bank.addInterest(account, 5.0);
         assertEquals(105.0, account.getBalance(), 0.001);
-    }
-
-    @Test
-    public void testTransactionHistoryAccessibleViaBank() {
-        Bank bank = new Bank();
-        Customer customer = new Customer("Alice");
-        bank.addCustomer(customer);
-        customer.getAccounts().get(0).deposit(200.0);
-        String history = bank.getCustomers().get(0).getAccounts().get(0).getTransactionHistory();
-        assertEquals("Deposited: 200.0\n", history);
-    }
-
-    @Test
-    public void testTransactionHistoryEmptyForNewAccount() {
-        Bank bank = new Bank();
-        Customer customer = new Customer("Bob");
-        bank.addCustomer(customer);
-        String history = bank.getCustomers().get(0).getAccounts().get(0).getTransactionHistory();
-        assertEquals("", history);
-    }
-
-    @Test
-    public void testTransactionHistoryAcrossMultipleCustomers() {
-        Bank bank = new Bank();
-        Customer alice = new Customer("Alice");
-        Customer bob = new Customer("Bob");
-        bank.addCustomer(alice);
-        bank.addCustomer(bob);
-        alice.getAccounts().get(0).deposit(100.0);
-        bob.getAccounts().get(0).deposit(500.0);
-        String aliceHistory = bank.getCustomers().get(0).getAccounts().get(0).getTransactionHistory();
-        String bobHistory = bank.getCustomers().get(1).getAccounts().get(0).getTransactionHistory();
-        assertEquals("Deposited: 100.0\n", aliceHistory);
-        assertEquals("Deposited: 500.0\n", bobHistory);
-    }
-
-    @Test
-    public void testFreezeAndUnfreezeAccount() {
-        Bank bank = new Bank();
-        BankAccount account = new BankAccount();
-
-        bank.freezeAccount(account);
-        assertTrue(account.isFrozen());
-
-        bank.unfreezeAccount(account);
-        assertFalse(account.isFrozen());
     }
 }
