+++ src/test/CustomerTest.java
@@ -4,8 +4,6 @@ import main.Customer;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 public class CustomerTest {
@@ -26,29 +24,6 @@ public class CustomerTest {
 
         try {
             customer.updatePersonalInformation(null, "5551234567", "Nick@test.com");
-    public void testSetPasswordAndVerifyPassword() {
-        Customer customer = new Customer("Nick");
-        customer.setPassword("securePass123");
-
-        assertTrue(customer.verifyPassword("securePass123"));
-        assertFalse(customer.verifyPassword("wrongPassword"));
-    }
-
-    @Test
-    public void testSetPinAndVerifyPin() {
-        Customer customer = new Customer("Nick");
-        customer.setPin("1234");
-
-        assertTrue(customer.verifyPin("1234"));
-        assertFalse(customer.verifyPin("5678"));
-    }
-
-    @Test
-    public void testSetPinRejectsInvalidPin() {
-        Customer customer = new Customer("Nick");
-
-        try {
-            customer.setPin("12a4");
             fail();
         } catch (IllegalArgumentException e) {
             // expected
